requirements.txt
# Python requirements file for fine-tuning language models
torch>=2.0.0
transformers>=4.34.0
datasets>=2.14.0
peft>=0.5.0
trl>=0.7.2
accelerate>=0.21.0
bitsandbytes>=0.41.0
sentencepiece
protobuf
scipy
evaluate
nats-py
pytest-asyncio
PyYAML>=6.0
